Introducing MVVM into your SwiftUI project
All the logic is handled outside the view, so later on when you learn to write tests you’ll find the view model is much easier to work with.
When we write data we’re making iOS encrypt it so the file can’t be read or written until the user unlocks their device.
The load and save process is almost transparent – we added one modifier and changed another, and that’s all it took.